+++ app/views/admin_incoming_message/_actions.html.erb
@@ -14,7 +14,7 @@
   <% end %>
 <% end %>
 
-<%= form_tag redeliver_admin_incoming_message_path(incoming_message), class: 'form form-inline' do %>
+<%= form_tag admin_incoming_message_redelivery_path(incoming_message), class: 'form form-inline' do %>
   <div class="control-group">
     <label class="control-label" for="url_title_<%= incoming_message.id %>">
       Redeliver message to one or more other requests
